refactor: split apart autosuggest inner UI logic with typeahead wrapper component
